moving ohms_footnotes.js out of /audio directory as it has nothing to do with audio. Also updating comments to better highlight the fact that we have two separate uses of bootstrap tooltips on the site now.
