+++ app/javascript/packs/application.js
@@ -22,11 +22,15 @@ import '../src/js/custom_google_analytics_events.js';
 import '../src/js/cart_control.js';
 import '../src/js/date_range_render_workaround.js';
 import '../src/js/tab_selection_in_anchor';
+
+
+// Generic tooltips
 import '../src/js/bootstrap_tooltips_activate';
+// and special OHMS footnotes tooltips
+import '../src/js/ohms_footnotes.js';
 
 import '../src/js/audio/play_at_timecode.js';
 import '../src/js/audio/ohms_search.js';
-import '../src/js/audio/ohms_footnotes.js';
 import '../src/js/audio/accordion_open_on_screen.js';
 import '../src/js/audio/navbar_tabs.js';
